refactor(security): centralize security constants in utils module
- Move CSRF configuration constants to SECURITY_CONFIG
- Extract sensitive field definitions to dedicated constants
- Consolidate security patterns and injection detection rules
- Add comprehensive file upload security constants
- Organize constants with proper documentation and categorization
