+++ utils/constants.util.js
@@ -50,7 +50,25 @@
 // - No third-party dependencies - pure Node.js implementation
 // - Compatible with CommonJS and ES6 module systems
 // - Environment-agnostic design works across all deployment targets
-//
+// =============================================================================
+
+// =============================================================================
+// CORE NODE.JS DEPENDENCIES
+// =============================================================================
+// No core Node.js dependencies required - pure constants implementation
+
+// =============================================================================
+// THIRD-PARTY DEPENDENCIES
+// =============================================================================
+// No third-party dependencies required
+
+// =============================================================================
+// INTERNAL DEPENDENCIES
+// =============================================================================
+// No internal dependencies - this is a foundational constants module
+
+// =============================================================================
+// APPLICATION ENVIRONMENT & MODES
 // =============================================================================
 
 /**
@@ -100,6 +118,10 @@ const MODES = {
  */
 const ROOT = process.cwd();
 
+// =============================================================================
+// FILE SYSTEM & DIRECTORY PATHS
+// =============================================================================
+
 /**
  * Application directory structure and file paths
  * @constant {Object}
@@ -119,7 +141,7 @@ const ROOT = process.cwd();
  * @property {string} DEBUG - Debug information and development artifacts
  * @property {string} DOCS - Documentation files and API specifications
  * @property {string} DOCS_PATHS - Path documentation and routing diagrams
- * @property {string} ROUTES_COMMON - Common route definitions and endpoint configurations
+ * @property {string} ROUTES_DEFAULT - Common route definitions and endpoint configurations
  *
  * @example
  * // Reading a template file
@@ -149,6 +171,10 @@ const PATHS = {
   DOCS_PATHS: ROOT + '/docs/paths',
   ROUTES_DEFAULT: ROOT + '/routes/web',
 };
+
+// =============================================================================
+// APPLICATION NAMESPACING & PREFIXES
+// =============================================================================
 
 /**
  * Application module prefixes for namespacing and routing
@@ -197,21 +223,9 @@ const PREFIXES = {
   USR: 'users',
 };
 
-/**
- * Supported client device types for user agent detection and responsive behavior
- * @constant {Array<string>}
- * @description Comprehensive list of device types used for feature targeting,
- *              user interface adaptation, and analytics tracking
- *
- * @example
- * // Device detection and feature flagging
- * if (DEVICES.includes(userDevice)) {
- *   enableDeviceSpecificFeatures(userDevice);
- * }
- *
- * @since Version 1.5.0
- */
-const DEVICES = ['common', 'web', 'app', 'wearable', 'desktop', 'bot'];
+// =============================================================================
+// NETWORK & HTTP CONSTANTS
+// =============================================================================
 
 /**
  * HTTP methods with numeric identifiers for efficient processing
@@ -250,6 +264,10 @@ const METHODS = {
   DELETE: 5,
   OPTIONS: 6,
 };
+
+// =============================================================================
+// DATA VALIDATION & PROCESSING
+// =============================================================================
 
 /**
  * Number handling, formatting, and validation constants
@@ -396,6 +414,10 @@ const ESCAPE_SEQUENCES = {
   '\\': '\\',
 };
 
+// =============================================================================
+// SECURITY & CRYPTOGRAPHY
+// =============================================================================
+
 /**
  * Cryptographic algorithms for encryption, hashing, and security operations
  * @constant {Object}
@@ -444,69 +466,6 @@ const KEY_SIZES = {
 };
 
 /**
- * Context keys for request metadata and application state management
- * @constant {Object}
- * @description Standardized keys for storing and accessing request context
- *              information throughout the application lifecycle
- * @property {string} USER_ID - Authenticated user identifier
- * @property {string} USER_DATA - Complete user profile and preferences
- * @property {string} SESSION_ID - User session identifier
- * @property {string} REQUEST_ID - Unique request identifier for tracing
- * @property {string} IP_ADDRESS - Client IP address for geo-location and rate limiting
- * @property {string} USER_AGENT - Client browser/device information
- * @property {string} TIMESTAMP - Request initiation timestamp
- * @property {string} PERMISSIONS - User permissions and access rights
- * @property {string} ROLES - User roles and group memberships
- * @property {string} TENANT_ID - Multi-tenant organization identifier
- * @property {string} CORRELATION_ID - Request correlation for distributed tracing
- * @property {string} TRANSACTION_ID - Financial or business transaction identifier
- * @property {string} LOCALE - User language and regional preferences
- * @property {string} TIMEZONE - User timezone for date/time formatting
- * @property {string} API_KEY - API authentication key
- * @property {string} CLIENT_ID - OAuth client identifier
- * @property {string} ORGANIZATION_ID - User organization identifier
- * @property {string} DEPARTMENT_ID - User department or team identifier
- * @property {string} CUSTOM_DATA - Extended custom context data
- * @property {string} SECURITY_CONTEXT - Security and authentication context
- * @property {string} AUDIT_CONTEXT - Audit trail and compliance data
- *
- * @example
- * // Setting request context
- * context.set(CONTEXT_KEYS.USER_ID, authenticatedUser.id);
- * context.set(CONTEXT_KEYS.REQUEST_ID, generateUUID());
- *
- * @example
- * // Accessing context in business logic
- * const userId = context.get(CONTEXT_KEYS.USER_ID);
- * const userRoles = context.get(CONTEXT_KEYS.ROLES);
- *
- * @since Version 1.6.0
- */
-const CONTEXT_KEYS = {
-  USER_ID: 'userId',
-  USER_DATA: 'userData',
-  SESSION_ID: 'sessionId',
-  REQUEST_ID: 'requestId',
-  IP_ADDRESS: 'ipAddress',
-  USER_AGENT: 'userAgent',
-  TIMESTAMP: 'timestamp',
-  PERMISSIONS: 'permissions',
-  ROLES: 'roles',
-  TENANT_ID: 'tenantId',
-  CORRELATION_ID: 'correlationId',
-  TRANSACTION_ID: 'transactionId',
-  LOCALE: 'locale',
-  TIMEZONE: 'timezone',
-  API_KEY: 'apiKey',
-  CLIENT_ID: 'clientId',
-  ORGANIZATION_ID: 'organizationId',
-  DEPARTMENT_ID: 'departmentId',
-  CUSTOM_DATA: 'customData',
-  SECURITY_CONTEXT: 'securityContext',
-  AUDIT_CONTEXT: 'auditContext',
-};
-
-/**
  * Comprehensive security configuration and policy definitions
  * @constant {Object}
  * @description Centralized security settings covering authentication,
@@ -531,7 +490,9 @@ const CONTEXT_KEYS = {
  * @property {number} SESSION.ABSOLUTE_TIMEOUT - Absolute session timeout (8 hours)
  * @property {Object} CSRF - CSRF protection configuration
  * @property {number} CSRF.TOKEN_LENGTH - CSRF token length (32 characters)
- * @property {number} CSRF.TOKEN_EXPIRY - CSRF token expiry (1 hour)
+ * @property {number} CSRF.TOKEN_TTL - CSRF token expiry (1 hour)
+ * @property {string} CSRF.HEADER_NAME - CSRF token header name
+ * @property {string} CSRF.COOKIE_NAME - CSRF token cookie name
  * @property {Object} VALIDATION - Input validation limits
  * @property {number} VALIDATION.MAX_STRING_LENGTH - Maximum string length (10000)
  * @property {number} VALIDATION.MAX_ARRAY_LENGTH - Maximum array size (1000)
@@ -579,7 +540,9 @@ const SECURITY_CONFIG = {
 
   CSRF: {
     TOKEN_LENGTH: 32,
-    TOKEN_EXPIRY: 60 * 60 * 1000, // 1 hour
+    TOKEN_TTL: 3600,
+    HEADER_NAME: 'x-csrf-token',
+    COOKIE_NAME: 'csrf-token',
   },
 
   VALIDATION: {
@@ -620,8 +583,15 @@ const SECURITY_CONFIG = {
  * @see {@link SECURITY_CONFIG} for related security settings
  */
 const SECURITY_PATTERNS = {
-  XSS: /<script\b[^<]*(?:(?!<\/script>)<[^<]*)*<\/script>/gi,
-  SQL_INJECTION: /((\%3D)|(=))[^\n]*((\%27)|(\')|(\-\-)|(\%3B)|(;))/i,
+  XSS: [/<script[^>]*>.*?<\/script>/gi, /on\w+\s*=/gi, /javascript:/gi, /<iframe/gi, /<object/gi, /<embed/gi],
+  SQL_INJECTION: [
+    /(\b(SELECT|INSERT|UPDATE|DELETE|DROP|CREATE|ALTER|EXEC|EXECUTE|UNION|DECLARE)\b)/gi,
+    /(--|\;|\*|\/\*|\*\/|xp_|sp_)/gi,
+    /('|('')|;|--|\/\*|\*\/|xp_|sp_)/gi,
+  ],
+  NOSQL_INJECTION_PATTERNS: [/\$where/gi, /\$ne/gi, /\$gt/gi, /\$lt/gi, /\$regex/gi, /\$or/gi, /\$and/gi],
+  COMMAND_INJECTION_PATTERNS: [/[;&|`$(){}[\]<>]/g, /\.\.\//g, /(curl|wget|nc|netcat|bash|sh|powershell|cmd)/gi],
+  PATH_TRAVERSAL_PATTERNS: [/\.\.[\/\\]/g, /[\/\\]\.\./g, /%2e%2e/gi, /%252e/gi],
   HTML_TAGS: /<[^>]*>/g,
   SPECIAL_CHARS: /[<>\"'%;()&+]/g,
   EMAIL: /^[^\s@]+@[^\s@]+\.[^\s@]+$/,
@@ -658,6 +628,10 @@ const THREAT_LEVELS = {
   HIGH: 'high',
   CRITICAL: 'critical',
 };
+
+// =============================================================================
+// PERFORMANCE & CACHING
+// =============================================================================
 
 /**
  * Caching configuration and performance optimization parameters
@@ -719,6 +693,10 @@ const PERFORMANCE_CONFIG = {
   MEMORY_WARNING_THRESHOLD: 0.85, // 85% of available memory
   CPU_WARNING_THRESHOLD: 0.8, // 80% CPU usage
 };
+
+// =============================================================================
+// DATABASE & STORAGE
+// =============================================================================
 
 /**
  * Database query configuration and search optimization parameters
@@ -790,6 +768,26 @@ const DB_CONFIG = {
   },
 };
 
+// =============================================================================
+// CLIENT & DEVICE MANAGEMENT
+// =============================================================================
+
+/**
+ * Supported client device types for user agent detection and responsive behavior
+ * @constant {Array<string>}
+ * @description Comprehensive list of device types used for feature targeting,
+ *              user interface adaptation, and analytics tracking
+ *
+ * @example
+ * // Device detection and feature flagging
+ * if (DEVICES.includes(userDevice)) {
+ *   enableDeviceSpecificFeatures(userDevice);
+ * }
+ *
+ * @since Version 1.5.0
+ */
+const DEVICES = ['common', 'web', 'app', 'wearable', 'desktop', 'bot'];
+
 /**
  * Client device type classifications for analytics and feature targeting
  * @constant {Object}
@@ -819,6 +817,10 @@ const DEVICE_TYPES = {
   DESKTOP_APP: 'desktop_app',
   GAME_CONSOLE: 'game_console',
 };
+
+// =============================================================================
+// LOGGING & MONITORING
+// =============================================================================
 
 /**
  * Log level definitions following syslog standard (RFC 5424)
@@ -890,6 +892,10 @@ const LOG_COLORS = {
   debug: 'blue',
 };
 
+// =============================================================================
+// DATA SENSITIVITY & PRIVACY
+// =============================================================================
+
 /**
  * Sensitive data field names for automatic masking and protection
  * @constant {Array<string>}
@@ -908,16 +914,200 @@ const LOG_COLORS = {
  */
 const SENSITIVE_FIELDS = [
   'password',
-  'token',
-  'authorization',
-  'cookie',
-  'session',
-  'creditcard',
-  'credit_card',
-  'cvv',
+  'passwordHash',
+  'salt',
+  'securityToken',
+  'resetToken',
+  'verificationToken',
+  'apiKey',
+  'apiSecret',
+  'privateKey',
+  'refreshToken',
+  'accessToken',
+  'sessionId',
+  'sessionToken',
   'ssn',
-  'secret',
+  'taxId',
+  'creditCard',
+  'cvv',
+  'pin',
+  'bankAccount',
 ];
+
+/**
+ * Protected system fields that should not be modified by users
+ * @constant {Array<string>}
+ * @description System-managed fields that are controlled internally
+ *              and should not be directly modified through user input
+ *
+ * @example
+ * // Filtering protected fields from user input
+ * const safeUpdate = omit(updateData, PROTECTED_FIELDS);
+ *
+ * @since Version 2.1.0
+ */
+const PROTECTED_FIELDS = [
+  'id',
+  'createdAt',
+  'updatedAt',
+  'deletedAt',
+  'password',
+  'rolId',
+  'securityLevel',
+  'isAdmin',
+  'isSuperAdmin',
+  'permissions',
+  'scopes',
+  'verified',
+  'emailConfirmedAt',
+  'mobileNumberConfirmedAt',
+];
+
+/**
+ * Role-based field protection for different user types
+ * @constant {Object}
+ * @description Defines which fields should be protected based on user roles,
+ *              enabling fine-grained access control for different user types
+ * @property {Array<string>} user - Fields protected for regular users
+ * @property {Array<string>} admin - Fields protected for administrators
+ * @property {Array<string>} system - Fields protected for system users
+ *
+ * @example
+ * // Role-based field filtering
+ * const allowedFields = getAccessibleFields(userRole, updateData);
+ * const safeUpdate = pick(updateData, allowedFields);
+ *
+ * @since Version 2.1.0
+ */
+const ROLE_PROTECTED_FIELDS = {
+  user: ['rolId', 'securityLevelId', 'permissions', 'scopes', 'isAdmin'],
+  admin: ['isSuperAdmin', 'systemRole'],
+  system: [],
+};
+
+/**
+ * Personally Identifiable Information fields for GDPR/Privacy compliance
+ * @constant {Array<string>}
+ * @description PII fields that require special handling under privacy regulations
+ *              like GDPR, CCPA, and other data protection laws
+ *
+ * @example
+ * // PII data anonymization
+ * const anonymizedData = anonymizePII(userData, PII_FIELDS);
+ *
+ * @since Version 2.1.0
+ */
+const PII_FIELDS = [
+  'email',
+  'mobileNumber',
+  'phoneNumber',
+  'address',
+  'fullAddress',
+  'zipCode',
+  'postalCode',
+  'dateOfBirth',
+  'birthDate',
+  'identificationNumber',
+  'passport',
+  'driverLicense',
+];
+
+/**
+ * Internal system fields for audit and operational purposes
+ * @constant {Array<string>}
+ * @description Internal operational fields used for system management,
+ *              audit trails, and internal tracking purposes
+ *
+ * @example
+ * // Filter internal fields from external responses
+ * const externalResponse = omit(internalData, INTERNAL_FIELDS);
+ *
+ * @since Version 2.1.0
+ */
+const INTERNAL_FIELDS = [
+  'deletedAt',
+  'internalCode',
+  'internalId',
+  'systemId',
+  'securityLevel',
+  'securityLevelId',
+  'rolId',
+  'employeeId',
+  'createdBy',
+  'updatedBy',
+  'deletedBy',
+];
+
+// =============================================================================
+// CONTEXT & REQUEST MANAGEMENT
+// =============================================================================
+
+/**
+ * Context keys for request metadata and application state management
+ * @constant {Object}
+ * @description Standardized keys for storing and accessing request context
+ *              information throughout the application lifecycle
+ * @property {string} USER_ID - Authenticated user identifier
+ * @property {string} USER_DATA - Complete user profile and preferences
+ * @property {string} SESSION_ID - User session identifier
+ * @property {string} REQUEST_ID - Unique request identifier for tracing
+ * @property {string} IP_ADDRESS - Client IP address for geo-location and rate limiting
+ * @property {string} USER_AGENT - Client browser/device information
+ * @property {string} TIMESTAMP - Request initiation timestamp
+ * @property {string} PERMISSIONS - User permissions and access rights
+ * @property {string} ROLES - User roles and group memberships
+ * @property {string} TENANT_ID - Multi-tenant organization identifier
+ * @property {string} CORRELATION_ID - Request correlation for distributed tracing
+ * @property {string} TRANSACTION_ID - Financial or business transaction identifier
+ * @property {string} LOCALE - User language and regional preferences
+ * @property {string} TIMEZONE - User timezone for date/time formatting
+ * @property {string} API_KEY - API authentication key
+ * @property {string} CLIENT_ID - OAuth client identifier
+ * @property {string} ORGANIZATION_ID - User organization identifier
+ * @property {string} DEPARTMENT_ID - User department or team identifier
+ * @property {string} CUSTOM_DATA - Extended custom context data
+ * @property {string} SECURITY_CONTEXT - Security and authentication context
+ * @property {string} AUDIT_CONTEXT - Audit trail and compliance data
+ *
+ * @example
+ * // Setting request context
+ * context.set(CONTEXT_KEYS.USER_ID, authenticatedUser.id);
+ * context.set(CONTEXT_KEYS.REQUEST_ID, generateUUID());
+ *
+ * @example
+ * // Accessing context in business logic
+ * const userId = context.get(CONTEXT_KEYS.USER_ID);
+ * const userRoles = context.get(CONTEXT_KEYS.ROLES);
+ *
+ * @since Version 1.6.0
+ */
+const CONTEXT_KEYS = {
+  USER_ID: 'userId',
+  USER_DATA: 'userData',
+  SESSION_ID: 'sessionId',
+  REQUEST_ID: 'requestId',
+  IP_ADDRESS: 'ipAddress',
+  USER_AGENT: 'userAgent',
+  TIMESTAMP: 'timestamp',
+  PERMISSIONS: 'permissions',
+  ROLES: 'roles',
+  TENANT_ID: 'tenantId',
+  CORRELATION_ID: 'correlationId',
+  TRANSACTION_ID: 'transactionId',
+  LOCALE: 'locale',
+  TIMEZONE: 'timezone',
+  API_KEY: 'apiKey',
+  CLIENT_ID: 'clientId',
+  ORGANIZATION_ID: 'organizationId',
+  DEPARTMENT_ID: 'departmentId',
+  CUSTOM_DATA: 'customData',
+  SECURITY_CONTEXT: 'securityContext',
+  AUDIT_CONTEXT: 'auditContext',
+};
+
+// =============================================================================
+// DEBUG & DEVELOPMENT
+// =============================================================================
 
 /**
  * Debug and development configuration parameters
@@ -941,6 +1131,10 @@ const DEBUG_SETTINGS = {
   DEFAULT_LINE_LENGTH: 150,
   DEVELOPMENT_MODE_VALUE: 2,
 };
+
+// =============================================================================
+// FILE HANDLING & STORAGE
+// =============================================================================
 
 /**
  * Amazon S3 storage configuration and operation parameters
@@ -970,32 +1164,185 @@ const S3_CONFIG = {
   MAX_KEYS_PER_REQUEST: 1000,
 };
 
+/**
+ * Allowed MIME types for secure file upload validation
+ * @constant {Object}
+ * @description Categorized MIME type validation for different file types,
+ *              ensuring only safe and expected file types are processed
+ * @property {Array<string>} image - Allowed image MIME types
+ * @property {Array<string>} document - Allowed document MIME types
+ * @property {Array<string>} spreadsheet - Allowed spreadsheet MIME types
+ * @property {Array<string>} video - Allowed video MIME types
+ * @property {Array<string>} audio - Allowed audio MIME types
+ *
+ * @example
+ * // MIME type validation
+ * if (ALLOWED_MIME_TYPES.image.includes(file.mimetype)) {
+ *   await processImageUpload(file);
+ * }
+ *
+ * @since Version 2.2.0
+ */
+const ALLOWED_MIME_TYPES = {
+  image: ['image/jpeg', 'image/png', 'image/gif', 'image/webp', 'image/svg+xml'],
+  document: [
+    'application/pdf',
+    'application/msword',
+    'application/vnd.openxmlformats-officedocument.wordprocessingml.document',
+  ],
+  spreadsheet: ['application/vnd.ms-excel', 'application/vnd.openxmlformats-officedocument.spreadsheetml.sheet'],
+  video: ['video/mp4', 'video/mpeg', 'video/quicktime', 'video/webm'],
+  audio: ['audio/mpeg', 'audio/wav', 'audio/ogg'],
+};
+
+/**
+ * Dangerous file extensions that should be blocked from upload
+ * @constant {Array<string>}
+ * @description Executable and potentially dangerous file extensions
+ *              that should be automatically rejected for security
+ *
+ * @example
+ * // File extension validation
+ * if (DANGEROUS_EXTENSIONS.some(ext => filename.endsWith(ext))) {
+ *   throw new Error('File type not allowed');
+ * }
+ *
+ * @since Version 2.2.0
+ */
+const DANGEROUS_EXTENSIONS = [
+  '.exe',
+  '.bat',
+  '.cmd',
+  '.com',
+  '.pif',
+  '.scr',
+  '.vbs',
+  '.js',
+  '.jar',
+  '.msi',
+  '.app',
+  '.deb',
+  '.rpm',
+  '.dmg',
+  '.pkg',
+  '.sh',
+  '.bash',
+  '.ps1',
+];
+
+/**
+ * File signature mappings for content-type validation
+ * @constant {Object}
+ * @description Hex signature to MIME type mappings for validating
+ *              file contents regardless of extension or declared type
+ * @property {string} ffd8ffe0 - JPEG image signature
+ * @property {string} ffd8ffe1 - JPEG image signature
+ * @property {string} ffd8ffe2 - JPEG image signature
+ * @property {string} 89504e47 - PNG image signature
+ * @property {string} 47494638 - GIF image signature
+ * @property {string} 52494646 - WEBP image signature
+ * @property {string} 25504446 - PDF document signature
+ * @property {string} 504b0304 - ZIP archive signature
+ * @property {string} 504b0506 - ZIP archive signature
+ * @property {string} 504b0708 - ZIP archive signature
+ *
+ * @example
+ * // File content validation
+ * const fileSignature = fileBuffer.toString('hex', 0, 4);
+ * const detectedType = FILE_SIGNATURE_MAP[fileSignature];
+ *
+ * @since Version 2.2.0
+ */
+const FILE_SIGNATURE_MAP = {
+  ffd8ffe0: 'image/jpeg',
+  ffd8ffe1: 'image/jpeg',
+  ffd8ffe2: 'image/jpeg',
+  '89504e47': 'image/png',
+  47494638: 'image/gif',
+  52494646: 'image/webp',
+  25504446: 'application/pdf',
+  '504b0304': 'application/zip',
+  '504b0506': 'application/zip',
+  '504b0708': 'application/zip',
+};
+
 // =============================================================================
 // MODULE EXPORTS
 // =============================================================================
+
+/**
+ * Application Constants Module
+ * @module utils/constants
+ * @description Centralized configuration management providing single source
+ *              of truth for all application constants, paths, and settings
+ *
+ * @example
+ * // Import specific constant groups
+ * const { PATHS, MODES, SECURITY_CONFIG } = require('~/utils/constants.util.js');
+ *
+ * @example
+ * // Import all constants
+ * const CONSTANTS = require('~/utils/constants.util.js');
+ *
+ * @since Version 1.0.0
+ */
 module.exports = {
+  // Application Environment & Modes
   ROOT,
-  PATHS,
   MODES,
+
+  // File System & Directory Paths
+  PATHS,
+
+  // Application Namespacing & Prefixes
   PREFIXES,
-  DEVICES,
+
+  // Network & HTTP Constants
   METHODS,
+
+  // Data Validation & Processing
   NUMBER_CONSTANTS,
   STRING_CONSTANTS,
   ESCAPE_SEQUENCES,
+
+  // Security & Cryptography
   ALGORITHMS,
   KEY_SIZES,
-  CONTEXT_KEYS,
   SECURITY_CONFIG,
   SECURITY_PATTERNS,
   THREAT_LEVELS,
+
+  // Performance & Caching
   CACHE_CONFIG,
   PERFORMANCE_CONFIG,
+
+  // Database & Storage
   DB_CONFIG,
+
+  // Client & Device Management
+  DEVICES,
   DEVICE_TYPES,
+
+  // Logging & Monitoring
   LOG_LEVELS,
   LOG_COLORS,
+
+  // Data Sensitivity & Privacy
   SENSITIVE_FIELDS,
+  PROTECTED_FIELDS,
+  ROLE_PROTECTED_FIELDS,
+  PII_FIELDS,
+  INTERNAL_FIELDS,
+
+  // Context & Request Management
+  CONTEXT_KEYS,
+
+  // Debug & Development
   DEBUG_SETTINGS,
+
+  // File Handling & Storage
   S3_CONFIG,
+  ALLOWED_MIME_TYPES,
+  DANGEROUS_EXTENSIONS,
+  FILE_SIGNATURE_MAP,
 };
